17-02-22 add ft_putnbr_unsigned_fd to libft

+++ ft_printf_p.c
@@ -6,7 +6,7 @@
 /*   By: hbourgeo <hbourgeo@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/17 09:05:48 by hbourgeo          #+#    #+#             */
-/*   Updated: 2022/02/18 08:34:18 by hbourgeo         ###   ########.fr       */
+/*   Updated: 2022/02/18 08:42:23 by hbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,13 @@ int ft_printf_p (va_list arg)
 	void	*add;
 
 	size = 0;
-	add = (unsigned long int)va_arg(arg, void *);
+	add = va_arg(arg, void *);
 	if (add == 0)
 	{
 		size += ft_putstr_fd("0x0", 1);
 		return(size);
 	}
 	size += ft_putstr_fd("0x", 1);
-	size += ft_putnbr_base_fd (add, "0123456789abcdef", 1);
+	size += ft_putnbr_base_fd ((unsigned long int)add, "0123456789abcdef", 1);
 	return(size);
 }
